refactor: extract shared TypeScript types to shared/ package
Move domain types (ApiType, WebDocument, SearchResult, ListItem),
constants (DEFAULT_API_URLS), and factory values (emptyDocument) from
web_interface_react into a shared/ directory. Both frontends now import
via @lenie/shared alias (tsconfig paths + Vite resolve.alias).

- Create shared/types/ with api.ts, documents.ts, and barrel index.ts
- Configure @lenie/shared alias in both web_interface_react and web_interface_app2
- Replace hardcoded DEFAULT_API_URL in app2 with import from shared
- Update Dockerfile and compose.yaml for new build context
- Add docs/shared-types.md and update CLAUDE.md files
