+++ shared/types/index.ts
@@ -1,0 +1,5 @@
+export type { ApiType } from './api';
+export { DEFAULT_API_URLS } from './api';
+
+export type { WebDocument, SearchResult, ListItem } from './documents';
+export { emptyDocument } from './documents';
