+++ web_interface_react/vite.config.ts
@@ -1,9 +1,15 @@
 /// <reference types="vitest" />
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
+import path from 'path';
 
 export default defineConfig({
   plugins: [react()],
+  resolve: {
+    alias: {
+      '@lenie/shared': path.resolve(__dirname, '../shared'),
+    },
+  },
   server: {
     port: 3000,
   },
